Add files via upload
This project was done in a VM, this was more of a "make sure you can do it in a VM" regarding the Linux command line configuration before actually performing Linux command line configurations on a real raspberry pi, so I had to create 2 documentations for both the DHCP & DNS server. So this is more of a project plan so I created documentation for it and then I submitted it to my fellow CSC members for review.