Add files via upload
This was a project I did out of curiosity, so in this project using the secure internal lab I did before now I have an environment in which I can test stuff I would like to try. So I tried a gray box pentest using the 2 VMs that are in the secure internal network. One VM is an attacker and the other one is the defender although both of these VMs run kali linux I just went with that. Feel free to check this project out!